+++ components/NewSpotForm/index.js
