+++ components/NewSpotForm/style.js
